Moved Response Helper classes around
