+++ src/test/java/org/openpnp/machine/photon/protocol/ResponsesHelperTest.java
@@ -3,6 +3,7 @@ package org.openpnp.machine.photon.protocol;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.openpnp.machine.photon.helpers.ResponsesHelper;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,15 +11,15 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 /**
  * These tests just exist to verify the TestResponsesHelper which is used in the PhotonFeedersTest
  */
-public class TestResponsesHelperTest {
+public class ResponsesHelperTest {
     protected static String uuid_s = "FFEEDDCCBBAA998877665544";
     protected static int[] uuid_b = {0xFF, 0xEE, 0xDD, 0xCC, 0xBB, 0xAA, 0x99, 0x88, 0x77, 0x66, 0x55, 0x44};
 
-    private TestResponsesHelper testResponses;
+    private ResponsesHelper testResponses;
 
     @BeforeEach
     public void setUp() {
-        testResponses = new TestResponsesHelper(0);
+        testResponses = new ResponsesHelper(0);
     }
